Update autosize.js
I tried your npm package in my website and i noticed that an x axis scroll bar appeared into my textbox. So i then passed through your code then met " ta.style.overflow = `scroll` " in line 65. I thought it is best to sepparate the overflow for the y and x axis so as to always hide the x axis scroll bar. 

If you think this is not a good solution I wont agree less, am new to programming and I trust you and the fellow contributors for creating this useful tool. Thanks in advance for taking your time to pass through this small change request.🤗